chore(auth): handle err and log
